requirements.txt
google_api_python_client==2.63.0
google_auth_oauthlib==0.5.3
pandas==1.3.4
protobuf==4.21.7
python_dateutil==2.8.2
selenium==4.5.0
webdriver_manager==3.8.3
